HTML Files from GCM Bootstrap Design
Exported the HTML files along with their assets to a new folder. Added the Google Translate button (horizontal) to the Homeapage.html.
Made a basic Webserver.py as I was running into issues with browsers trying to open the HTML file and running the Google Translate API. The scriptLang.js file is currently unused but will try to implement it as it's for the Google Translate bar to not appear after translation.